@@ -21,7 +21,9 @@
     "@babel/core": "^7.4.5",
     "@babel/plugin-transform-modules-commonjs": "^7.4.4",
     "camelcase": "^5.3.1",
-    "tslib": "^1.9.3"
+    "slash": "^3.0.0",
+    "tslib": "^1.9.3",
+    "typescript": "^3.5.1"
   },
   "peerDependencies": {
     "@orbit/data": "^0.15.23",
@@ -35,7 +37,6 @@
     "jest": "^24.7.1",
     "ts-jest": "^24.0.2",
     "tslint": "^5.16.0",
-    "tslint-config-standard": "^8.0.1",
-    "typescript": "^3.4.5"
+    "tslint-config-standard": "^8.0.1"
   }
 }
