@@ -1,9 +1,8 @@
 module.exports = {
-  'roots': [
-    '<rootDir>/src'
-  ],
-  'transform': {
+  roots: ['<rootDir>/src'],
+  transform: {
     '^.+\\.ts$': 'ts-jest'
   },
-  moduleFileExtensions: ['js', 'ts', 'json']
+  moduleFileExtensions: ['js', 'ts', 'json'],
+  testPathIgnorePatterns: ['/node_modules/', '/fixtures/']
 }
