@@ -1,12 +1,36 @@
+import path from 'path'
 import { Schema, SchemaSettings } from '@orbit/data'
 import { generateTypes } from '../generator'
 
 describe('generateTypes', () => {
-  it('should generate the generic types', async () => {
+  // Basic
+
+  it('should generate a header', async () => {
     const definition: SchemaSettings = {
       models: {}
     }
-    expect(generateTypes(new Schema(definition))).toContain(`GenericRecord`)
+    const types = generateTypes(new Schema(definition))
+    expect(types).toContain(`Record`)
+  })
+
+  it('should generate a user record interface', async () => {
+    const definition = {
+      models: {
+        user: {}
+      }
+    }
+    const types = generateTypes(new Schema(definition))
+    expect(types).toContain(`UserRecord extends Record, UserRecordIdentity`)
+  })
+
+  it('should generate a user identity interface', async () => {
+    const definition = {
+      models: {
+        user: {}
+      }
+    }
+    const types = generateTypes(new Schema(definition))
+    expect(types).toContain(`interface UserRecordIdentity`)
   })
 
   it('should generate attributes', async () => {
@@ -20,23 +44,8 @@ describe('generateTypes', () => {
       }
     }
     const types = generateTypes(new Schema(definition))
-    expect(types).toContain(`UserModel`)
     expect(types).toContain(`UserAttributes`)
     expect(types).toContain(`username: string`)
-    expect(types).not.toContain(`UserRelationships`)
-  })
-
-  it('should generate attributes as any if type if not set', async () => {
-    const definition = {
-      models: {
-        user: {
-          attributes: {
-            username: {}
-          }
-        }
-      }
-    }
-    expect(generateTypes(new Schema(definition))).toContain(`username: any`)
   })
 
   it('should generate relationships', async () => {
@@ -44,18 +53,14 @@ describe('generateTypes', () => {
       models: {
         user: {
           relationships: {
-            group: { type: 'hasOne' },
-            peers: { type: 'hasMany' }
+            group: { type: 'hasOne' }
           }
         }
       }
     } as const
     const types = generateTypes(new Schema(definition))
-    expect(types).toContain(`UserModel`)
     expect(types).toContain(`UserRelationships`)
     expect(types).toContain(`group: RecordHasOneRelationship`)
-    expect(types).toContain(`peers: RecordHasManyRelationship`)
-    expect(types).not.toContain(`UserAttributes`)
   })
 
   it('should generate both attributes and relationships', async () => {
@@ -72,11 +77,8 @@ describe('generateTypes', () => {
       }
     } as const
     const types = generateTypes(new Schema(definition))
-    expect(types).toContain(`UserModel`)
     expect(types).toContain(`UserAttributes`)
-    expect(types).toContain(`username: string`)
     expect(types).toContain(`UserRelationships`)
-    expect(types).toContain(`group: RecordHasOneRelationship`)
   })
 
   it('should generate neither attributes and relationships', async () => {
@@ -86,8 +88,94 @@ describe('generateTypes', () => {
       }
     } as const
     const types = generateTypes(new Schema(definition))
-    expect(types).toContain(`UserModel`)
     expect(types).not.toContain(`UserAttributes`)
     expect(types).not.toContain(`UserRelationships`)
+  })
+
+  it('should generate extra imports', async () => {
+    const definition = {
+      models: {}
+    }
+    const types = generateTypes(new Schema(definition), {
+      extraImports: [{ type: 'Type', modulePath: 'src' }]
+    })
+    expect(types).toContain(`import { Type } from "./src/index"`)
+  })
+
+  it('should generate extra imports from a different base directory', async () => {
+    const definition = {
+      models: {}
+    }
+    const types = generateTypes(new Schema(definition), {
+      basePath: path.resolve(__dirname, '..'),
+      extraImports: [{ type: 'Type', modulePath: '.' }]
+    })
+    expect(types).toContain(`import { Type } from "./index"`)
+  })
+
+  // Attribute cases
+
+  it('should generate attributes as any if type if not set or unknown', async () => {
+    const definition = {
+      models: {
+        user: {
+          attributes: {
+            username: {},
+            password: { type: 'foo' }
+          }
+        }
+      }
+    }
+    const types = generateTypes(new Schema(definition))
+    expect(types).toContain(`username: any`)
+    expect(types).toContain(`password: any`)
+  })
+
+  it('should generate typescript import for typed attributes', async () => {
+    const definition = {
+      models: {
+        user: {
+          attributes: {
+            permission: { type: 'string', ts: 'AttributeDefinition' }
+          }
+        }
+      }
+    }
+    const types = generateTypes(new Schema(definition))
+    expect(types).toContain(`import { AttributeDefinition } from "./src/types"`)
+    expect(types).toContain(`permission: AttributeDefinition`)
+  })
+
+  it('should generate typescript import for multiple typed attributes', async () => {
+    const definition = {
+      models: {
+        user: {
+          attributes: {
+            permission: { type: 'string', ts: 'AttributeDefinition' },
+            group: { type: 'string', ts: 'ModelDefinition' }
+          }
+        }
+      }
+    }
+    const types = generateTypes(new Schema(definition))
+    expect(types).toContain(
+      `import { AttributeDefinition, ModelDefinition } from "./src/types"`
+    )
+  })
+
+  it('should generate typescript import for typed attributes from a different base directory', async () => {
+    const definition = {
+      models: {
+        user: {
+          attributes: {
+            permission: { type: 'string', ts: 'AttributeDefinition' }
+          }
+        }
+      }
+    }
+    const types = generateTypes(new Schema(definition), {
+      basePath: path.resolve(__dirname, '..')
+    })
+    expect(types).toContain(`import { AttributeDefinition } from "./types"`)
   })
 })
